Create and then take quizzes
QuizMaker and QuizTaker use Python and Python IDLE to create quizzes
that peopel can take. QuizMaker accepts input from the user to create
quizzes with multiple choice, fill-in-the-blanks, and true-or-false
questions. It then saves the input in a specific format in a text file,
which QuizTaker can open and then read.